+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node loader.js",
+    "start": "node index.js",
     "dev": "nodemon index.js"
   },
   "author": "",
