+++ src/index.html
@@ -2,6 +2,7 @@
 <html>
 
 <head>
+    <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
     <link rel="stylesheet" type="text/css" href="./index.css">
     <link rel="stylesheet" type="text/css" href="./checkmark.css">
 
@@ -10,6 +11,10 @@
 
 <body>
     <div id='root'></div>
+    <script src="./controller.js"></script>
+    <script src="./model.js"></script>
+    <script src="./view.js"></script>
+    <script src="./schemas.js"></script>
     <script src="./index.js"></script>
 </body>
